pyproject.toml
[project]
name = "python-project"
version = "0.1.0"
description = "python game"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.1",
]
